// src/features/analytics/beaconing/models/tls-tail.model.ts
import { z } from 'zod';

export const tlsTailSchema = z.object({
  tenant: z.number(),
  '@version': z.string(),
  proto: z.string(),
  agent: z.object({
    type: z.string(),
    name: z.string(),
    version: z.string(),
    ephemeral_id: z.string(),
    hostname: z.string(),
    id: z.string(),
  }),
  community_id: z.string(),
  see_id: z.string(),
  input: z.object({
    type: z.string(),
  }),
  ether: z.object({
    src_mac: z.string(),
    dest_mac: z.string(),
  }),
  tags: z.array(z.string()),
  type: z.string(),
  pkt_src: z.string(),
  hostname_info: z.object({
    subdomain: z.string(),
    tld: z.string(),
    domain: z.string(),
    domain_without_tld: z.string(),
    url: z.string(),
    host: z.string(),
  }),
  alerted: z.boolean().optional(),
  geoip: z.object({
    provider: z.object({
      autonomous_system_organization: z.string(),
      autonomous_system_number: z.number(),
    }),
    coordinate: z.array(z.number()),
    country_name: z.string(),
    longitude: z.number(),
    country: z.object({
      name: z.string(),
      geoname_id: z.number(),
      iso_code: z.string(),
    }),
    timezone: z.string(),
    country_code2: z.string(),
    location: z.object({
      lat: z.number(),
      lon: z.number(),
    }),
    ip: z.string(),
    latitude: z.number(),
    continent_code: z.string(),
    continent: z.object({
      code: z.string(),
      name: z.string(),
      geoname_id: z.number(),
    }),
    country_code3: z.string(),
    registered_country: z.object({
      name: z.string(),
      geoname_id: z.number(),
      iso_code: z.string(),
    }),
  }),
  src_ip: z.string(),
  app_proto: z.string(),
  tls: z.object({
    cipher_security: z.string(),
    notbefore: z.string(),
    ja3: z.object({
      string: z.string(),
      agent: z.array(z.string()),
      agent_count: z.number(),
      hash: z.string(),
    }),
    subject: z.string(),
    version: z.string(),
    notafter: z.string(),
    ja4: z.string(),
    issuerdn: z.string(),
    serial: z.string(),
    ja3s: z.object({
      string: z.string(),
      hash: z.string(),
    }),
    cipher_suite: z.string(),
    sni: z.string(),
    fingerprint: z.string(),
  }),
  dest_ip: z.string(),
  log: z.object({
    offset: z.number(),
    file: z.object({
      path: z.string(),
    }),
  }),
  host: z.string(),
  event_type: z.string(),
  dest_port: z.number(),
  '@timestamp': z.string(),
  src_port: z.number(),
  stamus_infrequent: z.boolean().optional(),
  flow_id: z.number(),
  net_info: z.object({
    src: z.array(z.string()),
    src_agg: z.string(),
  }),
  logger: z.string(),
  timestamp: z.string(),
  in_iface: z.string(),
  see_name: z.string(),
  _id: z.string(),
});

export type TlsTail = z.infer<typeof tlsTailSchema>;